+++ scripts/build.js
@@ -40,6 +40,11 @@ const languages = [
     dir: '',
     wasm: 'rust',
   },
+  {
+    npm: 'tree-sitter-go',
+    dir: '',
+    wasm: 'go',
+  },
 ]
 
 // Build wasm parsers for supported languages
@@ -53,6 +58,13 @@ if (!fs.existsSync(treeSitterCli)) {
 } else {
   for (const { npm, dir, wasm } of languages) {
     const module = path.join('node_modules', npm, dir)
+
+    if (!fs.existsSync(module)) {
+      console.error(
+        `Module ${module} does not exist. This is likely due to an installation and/or build failure of the module. Please check the logs.`
+      )
+      process.exit(1)
+    }
 
     let command
     if (process.env.CI) {